requirements.txt
flask==3.1.1
gunicorn==23.0.0
apscheduler==3.11.0
yfinance==0.2.54
pandas==2.2.3
numpy==2.2.3
requests==2.32.3
smartapi-python==1.3.6
pyotp==2.9.0
flask 
apscheduler